created updates the blog code to have correct redirects

+++ blog/views.py
@@ -2,7 +2,10 @@ from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.utils import timezone
 from .models import Blog
 from .forms import BlogPostForm
+from profiles.models import User_Profile
 from django.contrib import messages
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 
 
 def get_blogs(request):
@@ -25,8 +28,13 @@ def blog_detail(request, pk):
     return render(request, 'blog/blogdetail.html', context)
 
 
-def create_or_edit_blog(request, pk=None):
+@login_required
+def createblog_or_editblog_blog(request, pk=None):
     """ A view to createing and editing blogs """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admins can create blogs.')
+        return redirect(reverse('home'))
+
     blog = get_object_or_404(Blog, pk=pk) if pk else None
     if request.method == 'POST':
         form = BlogPostForm(request.POST, request.FILES, instance=blog)
@@ -43,8 +51,13 @@ def create_or_edit_blog(request, pk=None):
     return render(request, 'blog/blogpostform.html', context)
 
 
+@login_required
 def blog_delete(request, pk):
     """ A view to delete products in admin """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can delete blogs.')
+        return redirect(reverse('home'))
+
     blog = get_object_or_404(Blog, pk=pk)
     blog.delete()
     messages.warning(request, f'Blog Post { blog.title } was deleted!')
